feat: explore category in landing page, improve UX in search job in landing page
