+++ src/components/landing/job-categories.tsx
@@ -1,3 +1,6 @@
+'use client';
+
+import React, { useEffect, useState } from 'react';
 import {
   ArrowRight,
   Code,
@@ -10,66 +13,82 @@ import {
   Wrench,
 } from 'lucide-react';
 import Link from 'next/link';
+import { categoryService } from '@/services/api/category/category-api';
+import { Category } from '@/types/Category';
+
+// Icon mapping for categories
+const getCategoryIcon = (categoryName: string, isActive: boolean = false) => {
+  const iconClass = isActive ? 'h-6 w-6 text-white' : 'h-6 w-6 text-blue-600';
+
+  const iconMap: Record<string, React.ReactElement> = {
+    Design: <Wrench className={iconClass} />,
+    Sales: <Clock className={iconClass} />,
+    Marketing: <Megaphone className={iconClass} />,
+    Finance: <Wallet className={iconClass} />,
+    Technology: <Computer className={iconClass} />,
+    Engineering: <Code className={iconClass} />,
+    Business: <Briefcase className={iconClass} />,
+    'Human Resource': <Users className={iconClass} />,
+  };
+
+  return iconMap[categoryName] || <Briefcase className={iconClass} />;
+};
 
 export default function JobCategories() {
-  const categories = [
-    {
-      name: 'Design',
-      icon: <Wrench className="h-6 w-6 text-blue-600" />,
-      jobs: 235,
-      href: '#',
-      active: false,
-    },
-    {
-      name: 'Sales',
-      icon: <Clock className="h-6 w-6 text-blue-600" />,
-      jobs: 756,
-      href: '#',
-      active: false,
-    },
-    {
-      name: 'Marketing',
-      icon: <Megaphone className="h-6 w-6 text-white" />,
-      jobs: 140,
-      href: '#',
-      active: true,
-    },
-    {
-      name: 'Finance',
-      icon: <Wallet className="h-6 w-6 text-blue-600" />,
-      jobs: 325,
-      href: '#',
-      active: false,
-    },
-    {
-      name: 'Technology',
-      icon: <Computer className="h-6 w-6 text-blue-600" />,
-      jobs: 436,
-      href: '#',
-      active: false,
-    },
-    {
-      name: 'Engineering',
-      icon: <Code className="h-6 w-6 text-blue-600" />,
-      jobs: 542,
-      href: '#',
-      active: false,
-    },
-    {
-      name: 'Business',
-      icon: <Briefcase className="h-6 w-6 text-blue-600" />,
-      jobs: 211,
-      href: '#',
-      active: false,
-    },
-    {
-      name: 'Human Resource',
-      icon: <Users className="h-6 w-6 text-blue-600" />,
-      jobs: 346,
-      href: '#',
-      active: false,
-    },
-  ];
+  const [categories, setCategories] = useState<Category[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    const fetchCategories = async () => {
+      try {
+        setIsLoading(true);
+        setError(null);
+        const response = await categoryService.findAll();
+        setCategories(response);
+      } catch (err) {
+        console.error('Error fetching categories:', err);
+        setError('Failed to load categories');
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchCategories();
+  }, []);
+
+  // Loading skeleton
+  const LoadingSkeleton = () => (
+    <div className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-4">
+      {Array.from({ length: 8 }).map((_, index) => (
+        <div key={index} className="animate-pulse">
+          <div className="block rounded-lg border-1 bg-gray-200 p-6">
+            <div className="mb-4 h-6 w-6 rounded bg-gray-300"></div>
+            <div className="mb-2 h-6 w-3/4 rounded bg-gray-300"></div>
+            <div className="h-4 w-1/2 rounded bg-gray-300"></div>
+          </div>
+        </div>
+      ))}
+    </div>
+  );
+
+  // Error state
+  if (error) {
+    return (
+      <div className="bg-[#ffffff] px-4 py-12 md:px-8 lg:px-12">
+        <div className="mx-auto max-w-7xl">
+          <div className="mb-8 flex items-center justify-between">
+            <h2 className="text-4xl font-bold text-gray-800">
+              Explore by <span className="text-blue-500">category</span>
+            </h2>
+          </div>
+          <div className="rounded-lg border border-red-200 bg-red-50 p-6 text-center">
+            <p className="text-red-600">{error}</p>
+          </div>
+        </div>
+      </div>
+    );
+  }
 
   return (
     <div className="bg-[#ffffff] px-4 py-12 md:px-8 lg:px-12">
@@ -79,7 +98,7 @@ export default function JobCategories() {
             Explore by <span className="text-blue-500">category</span>
           </h2>
           <Link
-            href="#"
+            href="/job"
             className="group flex items-center gap-2 text-blue-600 hover:text-blue-700"
           >
             Show all jobs
@@ -87,34 +106,46 @@ export default function JobCategories() {
           </Link>
         </div>
 
-        <div className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-4">
-          {categories.map((category) => (
-            <Link
-              key={category.name}
-              href={category.href}
-              className={`block rounded-lg border-1 p-6 transition-transform hover:scale-[1.02] ${
-                category.active ? 'bg-blue-600' : 'bg-white'
-              }`}
-            >
-              <div className="mb-4">{category.icon}</div>
-              <h3
-                className={`mb-2 text-xl font-bold ${category.active ? 'text-white' : 'text-gray-900'}`}
+        {isLoading ? (
+          <LoadingSkeleton />
+        ) : (
+          <div className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-4">
+            {categories.map((category, index) => (
+              <Link
+                key={category.id}
+                href={`/job?categoryIds=${category.id}`}
+                className={`block rounded-lg border-1 p-6 transition-transform hover:scale-[1.02] ${
+                  index === 2 ? 'bg-blue-600' : 'bg-white'
+                }`}
               >
-                {category.name}
-              </h3>
-              <div className="flex items-center justify-between">
-                <span
-                  className={`${category.active ? 'text-blue-100' : 'text-gray-500'}`}
+                <div className="mb-4">
+                  {getCategoryIcon(category.name, index === 2)}
+                </div>
+                <h3
+                  className={`mb-2 text-xl font-bold ${
+                    index === 2 ? 'text-white' : 'text-gray-900'
+                  }`}
                 >
-                  {category.jobs} jobs available
-                </span>
-                <ArrowRight
-                  className={`h-4 w-4 ${category.active ? 'text-white' : 'text-gray-500'}`}
-                />
-              </div>
-            </Link>
-          ))}
-        </div>
+                  {category.name}
+                </h3>
+                <div className="flex items-center justify-between">
+                  <span
+                    className={`${
+                      index === 2 ? 'text-blue-100' : 'text-gray-500'
+                    }`}
+                  >
+                    Browse jobs
+                  </span>
+                  <ArrowRight
+                    className={`h-4 w-4 ${
+                      index === 2 ? 'text-white' : 'text-gray-500'
+                    }`}
+                  />
+                </div>
+              </Link>
+            ))}
+          </div>
+        )}
       </div>
     </div>
   );
